docker-image/web-scraper/requirements.txt
kafka-python>=2.0.2
lxml>=4.6.2
requests>=2.25.1
beautifulsoup4>=4.9.3